requirements.txt
flask==3.1.1
flask-cors==5.0.1
PyPDF2==3.0.1
pdfplumber==0.11.6
python-docx==1.1.2
spacy==3.8.4
scikit-learn==1.6.1
nltk==3.9.1
Werkzeug==3.1.3
pytesseract==0.3.13
Pillow==11.1.0
easyocr==1.7.2
